// src/lib/api/type.ts
import { AxiosRequestConfig, AxiosResponse } from "axios";

export type CustomConfigMeta = {
  appendSlash?: boolean;
  shouldResolve?: boolean;
  appendToken?: boolean; // Todo: implement this
  getAuthHeader?: () => Record<string, string> | undefined; // () => ({ Authorization: `Token ${getToken()}` }),
  formatErrorMessage?: () => any;
  errorResolver?: (error: any) => Error;
  responseResolver?: (response: AxiosResponse, meta: CustomConfigMeta) => any;
};

export interface AxiosWrapperConfig extends AxiosRequestConfig {
  meta?: CustomConfigMeta;
}